+++ Sources/AppUI/AppCommands.swift
@@ -5,6 +5,7 @@ public extension Notification.Name {
     static let aviRefreshRepository = Notification.Name("avi.refreshRepository")
     static let aviStageAll = Notification.Name("avi.stageAll")
     static let aviUnstageAll = Notification.Name("avi.unstageAll")
+    static let aviDiscardSelection = Notification.Name("avi.discardSelection")
     static let aviCommit = Notification.Name("avi.commit")
     static let aviFetchRepository = Notification.Name("avi.fetchRepository")
     static let aviPullRepository = Notification.Name("avi.pullRepository")